+++ keanu-python/docs/conf.py
@@ -23,9 +23,9 @@ copyright = u'2018, Improbable'
 author = u'Improbable'
 
 # The short X.Y version
-version = u'0.0.23'
+version = u'0.0.24'
 # The full version, including alpha/beta/rc tags
-release = u'0.0.23'
+release = u'0.0.24'
 
 # -- General configuration ---------------------------------------------------