+++ release_notes.md
@@ -1,3 +1,5 @@
+## Version 0.0.24
+
 ## Version 0.0.23
 
 ### Common